Added code to Scheduler and ElevatorEvent classes
I commented out the "numberOfFloors" variable in the Elevator class for now. The number of floors is also needed by the Scheduler to maintain its floor queues, so it would be bad to have this data stored in the Elevator class. For now I created a constant in the "Main" class but we could potentially create a separate class for setup related stuff (e.g. num floors, input file, etc.).